Add options for custom input types

surveyOptions() is a helper function for extendInputType. When defining custom input types, the choices arguments for shiny UI components should equal surveyOption(). See examples for more details.

surveyOptions()

Value

NA; used for side effects with extendInputType.

See also

Examples

extendInputType("inlineRadioButtons", {
shiny::radioButtons(
  inputId = surveyID(),
  label = surveyLabel(),
  selected = character(0),
  choices = surveyOptions(),
  inline = TRUE
)
})
#> Input Type "inlineRadioButtons" registered with {shinysurveys}. If the session restarts, you will need to re-register it.
#> To see all registered input extensions, please call `shinysurveys::listInputExtensions()`.